refactor(nav): extract sidebar nav data into a leaf module
Move NavItem, NavGroup, navigationGroups, and SUPPLEMENTAL_NAV_ITEMS out of
AppSidebar.nav.ts into a new AppSidebar.nav.data.ts that imports nothing
from @/lib/permissions. This is what will let areas.ts derive
AREA_DEFINITIONS from navigationGroups without an import cycle (nav.ts ->
routeAreas.ts -> areas.ts -> nav.ts would otherwise be a TDZ crash at
import time).

AppSidebar.nav.ts re-exports NavItem/NavGroup/navigationGroups so existing
importers (AppSidebar.tsx, permissions/preview.ts, tests) are unaffected.
Verbatim move, no behavior change.

Task 1 Step 1 of docs/superpowers/plans/2026-08-05-permissions-menu-mirror-pr1.md.
